Fixed tests and circular import
signer/remote and signer/universal now do not have circular import problems and
can use the functions from api/sign directly.
