+++ signer/remote/remote_test.go
@@ -2,7 +2,6 @@ package remote
 
 import (
 	"bytes"
-	"encoding/json"
 	"fmt"
 	"io/ioutil"
 	"math/big"
@@ -11,13 +10,12 @@ import (
 	"testing"
 	"time"
 
-	"github.com/cloudflare/cfssl/api"
 	apiinfo "github.com/cloudflare/cfssl/api/info"
+	apisign "github.com/cloudflare/cfssl/api/signhandler"
 	"github.com/cloudflare/cfssl/config"
-	"github.com/cloudflare/cfssl/errors"
 	"github.com/cloudflare/cfssl/helpers"
+	"github.com/cloudflare/cfssl/helpers/testsuite"
 	"github.com/cloudflare/cfssl/info"
-	"github.com/cloudflare/cfssl/log"
 	"github.com/cloudflare/cfssl/signer"
 	"github.com/cloudflare/cfssl/signer/local"
 )
@@ -59,7 +57,7 @@ var validMinimalAuthRemoteConfig = `
 }`
 
 func TestNewSigner(t *testing.T) {
-	remoteConfig := newConfig(t, []byte(validMinimalRemoteConfig))
+	remoteConfig := testsuite.NewConfig(t, []byte(validMinimalRemoteConfig))
 
 	_, err := NewSigner(remoteConfig.Signing)
 	if err != nil {
@@ -68,7 +66,7 @@ func TestNewSigner(t *testing.T) {
 }
 
 func TestNewAuthSigner(t *testing.T) {
-	remoteAuthConfig := newConfig(t, []byte(validMinimalAuthRemoteConfig))
+	remoteAuthConfig := testsuite.NewConfig(t, []byte(validMinimalAuthRemoteConfig))
 
 	_, err := NewSigner(remoteAuthConfig.Signing)
 	if err != nil {
@@ -80,7 +78,7 @@ func TestRemoteInfo(t *testing.T) {
 	remoteServer := newTestInfoServer(t)
 	defer closeTestServer(t, remoteServer)
 
-	remoteConfig := newConfig(t, []byte(validMinimalRemoteConfig))
+	remoteConfig := testsuite.NewConfig(t, []byte(validMinimalRemoteConfig))
 	// override with test server address, ignore url prefix "http://"
 	remoteConfig.Signing.OverrideRemotes(remoteServer.URL[7:])
 	s := newRemoteSigner(t, remoteConfig.Signing)
@@ -101,79 +99,32 @@ func TestRemoteInfo(t *testing.T) {
 	}
 }
 
-type csrTest struct {
-	file    string
-	keyAlgo string
-	keyLen  int
-	// Error checking function
-	errorCallback func(*testing.T, error)
-}
-
-var csrTests = []csrTest{
-	{
-		file:          "../local/testdata/rsa2048.csr",
-		keyAlgo:       "rsa",
-		keyLen:        2048,
-		errorCallback: nil,
-	},
-	{
-		file:          "../local/testdata/rsa3072.csr",
-		keyAlgo:       "rsa",
-		keyLen:        3072,
-		errorCallback: nil,
-	},
-	{
-		file:          "../local/testdata/rsa4096.csr",
-		keyAlgo:       "rsa",
-		keyLen:        4096,
-		errorCallback: nil,
-	},
-	{
-		file:          "../local/testdata/ecdsa256.csr",
-		keyAlgo:       "ecdsa",
-		keyLen:        256,
-		errorCallback: nil,
-	},
-	{
-		file:          "../local/testdata/ecdsa384.csr",
-		keyAlgo:       "ecdsa",
-		keyLen:        384,
-		errorCallback: nil,
-	},
-	{
-		file:          "../local/testdata/ecdsa521.csr",
-		keyAlgo:       "ecdsa",
-		keyLen:        521,
-		errorCallback: nil,
-	},
-}
-
 func TestRemoteSign(t *testing.T) {
 	remoteServer := newTestSignServer(t)
 	defer closeTestServer(t, remoteServer)
 
-	remoteConfig := newConfig(t, []byte(validMinimalRemoteConfig))
+	remoteConfig := testsuite.NewConfig(t, []byte(validMinimalRemoteConfig))
 	// override with test server address, ignore url prefix "http://"
 	remoteConfig.Signing.OverrideRemotes(remoteServer.URL[7:])
 	s := newRemoteSigner(t, remoteConfig.Signing)
 
 	hosts := []string{"cloudflare.com"}
-	for _, test := range csrTests {
-		csr, err := ioutil.ReadFile(test.file)
+	for _, test := range testsuite.CSRTests {
+		csr, err := ioutil.ReadFile(test.File)
 		if err != nil {
 			t.Fatal("CSR loading error:", err)
 		}
 		testSerial := big.NewInt(0x7007F)
 		certBytes, err := s.Sign(signer.SignRequest{
-			Hosts: hosts,
+			Hosts:   hosts,
 			Request: string(csr),
-			Serial: testSerial,
+			Serial:  testSerial,
 		})
-		if test.errorCallback != nil {
-			test.errorCallback(t, err)
+		if test.ErrorCallback != nil {
+			test.ErrorCallback(t, err)
 		} else {
 			if err != nil {
-				t.Fatalf("Expected no error. Got %s. Param %s %d", err.Error(), test.keyAlgo, test.keyLen)
+				t.Fatalf("Expected no error. Got %s. Param %s %d", err.Error(), test.KeyAlgo, test.KeyLen)
 			}
 			cert, err := helpers.ParseCertificatePEM(certBytes)
 			if err != nil {
@@ -192,7 +143,7 @@ func TestRemoteSignBadServerAndOverride(t *testing.T) {
 	defer closeTestServer(t, remoteServer)
 
 	// remoteConfig contains port 80 that no test server will listen on
-	remoteConfig := newConfig(t, []byte(validMinimalRemoteConfig))
+	remoteConfig := testsuite.NewConfig(t, []byte(validMinimalRemoteConfig))
 	s := newRemoteSigner(t, remoteConfig.Signing)
 
 	hosts := []string{"cloudflare.com"}
@@ -209,9 +160,9 @@ func TestRemoteSignBadServerAndOverride(t *testing.T) {
 	remoteConfig.Signing.OverrideRemotes(remoteServer.URL[7:])
 	s.SetPolicy(remoteConfig.Signing)
 	certBytes, err := s.Sign(signer.SignRequest{
-		Hosts: hosts,
+		Hosts:   hosts,
 		Request: string(csr),
-		Serial: big.NewInt(1),
+		Serial:  big.NewInt(1),
 	})
 	if err != nil {
 		t.Fatalf("Expected no error. Got %s.", err.Error())
@@ -224,17 +175,6 @@ func TestRemoteSignBadServerAndOverride(t *testing.T) {
 }
 
 // helper functions
-func newConfig(t *testing.T, configBytes []byte) *config.Config {
-	conf, err := config.LoadConfig([]byte(configBytes))
-	if err != nil {
-		t.Fatal("config loading error:", err)
-	}
-	if !conf.Valid() {
-		t.Fatal("config is not valid")
-	}
-	return conf
-}
-
 func newRemoteSigner(t *testing.T, policy *config.Signing) *Signer {
 	s, err := NewSigner(policy)
 	if err != nil {
@@ -310,107 +250,11 @@ func newHandler(t *testing.T, caFile, caKeyFile, op string) (http.Handler, error
 		t.Fatal(err)
 	}
 	if op == "sign" {
-		return NewSignHandlerFromSigner(s)
+		return apisign.NewHandlerFromSigner(s)
 	} else if op == "info" {
 		return apiinfo.NewHandler(s)
 	}
 
 	t.Fatal("Bad op code")
 	return nil, nil
-}
-
-// NewSignHandlerFromSigner generates a new SignHandler directly from
-// an existing signer.
-func NewSignHandlerFromSigner(s signer.Signer) (h http.Handler, err error) {
-	policy := s.Policy()
-	if policy == nil {
-		err = errors.New(errors.PolicyError, errors.InvalidPolicy)
-		return
-	}
-
-	// Sign will only respond for profiles that have no auth provider.
-	// So if all of the profiles require authentication, we return an error.
-	haveUnauth := (policy.Default.Provider == nil)
-	for _, profile := range policy.Profiles {
-		if !haveUnauth {
-			break
-		}
-		haveUnauth = (profile.Provider == nil)
-	}
-
-	if !haveUnauth {
-		err = errors.New(errors.PolicyError, errors.InvalidPolicy)
-		return
-	}
-
-	return &api.HTTPHandler{
-		Handler: &SignHandler{
-			signer: s,
-		},
-		Methods: []string{"POST"},
-	}, nil
-}
-
-// A SignHandler accepts requests with a hostname and certficate
-// parameter (which should be PEM-encoded) and returns a new signed
-// certificate. It includes upstream servers indexed by their
-// profile name.
-type SignHandler struct {
-	signer signer.Signer
-}
-
-// Handle responds to requests for the CA to sign the certificate request
-// present in the "certificate_request" parameter for the host named
-// in the "hostname" parameter. The certificate should be PEM-encoded. If
-// provided, subject information from the "subject" parameter will be used
-// in place of the subject information from the CSR.
-func (h *SignHandler) Handle(w http.ResponseWriter, r *http.Request) error {
-	log.Info("signature request received")
-
-	body, err := ioutil.ReadAll(r.Body)
-	if err != nil {
-		return err
-	}
-	r.Body.Close()
-
-	var req signer.SignRequest
-	err = json.Unmarshal(body, &req)
-	if err != nil {
-		return err
-	}
-
-	if len(req.Hosts) == 0 {
-		return errors.NewBadRequestString("missing paratmeter 'hosts'")
-	}
-
-	if req.Request == "" {
-		return errors.NewBadRequestString("missing parameter 'certificate_request'")
-	}
-
-	var cert []byte
-	var profile *config.SigningProfile
-
-	policy := h.signer.Policy()
-	if policy != nil && policy.Profiles != nil && req.Profile != "" {
-		profile = policy.Profiles[req.Profile]
-	}
-
-	if profile == nil && policy != nil {
-		profile = policy.Default
-	}
-
-	if profile.Provider != nil {
-		log.Error("profile requires authentication")
-		return errors.NewBadRequestString("authentication required")
-	}
-
-	cert, err = h.signer.Sign(req)
-	if err != nil {
-		log.Warningf("failed to sign request: %v", err)
-		return err
-	}
-
-	result := map[string]string{"certificate": string(cert)}
-	log.Info("wrote response")
-	return api.SendResponse(w, result)
 }
