+++ signer/universal/universal_test.go
@@ -2,7 +2,6 @@ package universal
 
 import (
 	"bytes"
-	"encoding/json"
 	"io/ioutil"
 	"math/big"
 	"net/http"
@@ -11,13 +10,12 @@ import (
 	"testing"
 	"time"
 
-	"github.com/cloudflare/cfssl/api"
 	apiinfo "github.com/cloudflare/cfssl/api/info"
+	apisign "github.com/cloudflare/cfssl/api/signhandler"
 	"github.com/cloudflare/cfssl/config"
-	"github.com/cloudflare/cfssl/errors"
 	"github.com/cloudflare/cfssl/helpers"
+	"github.com/cloudflare/cfssl/helpers/testsuite"
 	"github.com/cloudflare/cfssl/info"
-	"github.com/cloudflare/cfssl/log"
 	"github.com/cloudflare/cfssl/signer"
 )
 
@@ -49,7 +47,7 @@ var validMinimalRemoteConfig = `
       "CA": {
         "usages": [
           "cert sign",
-          "clr sign"
+          "crl sign"
         ],
         "expiry": "720h",
         "auth_key": "ca-auth"
@@ -78,7 +76,7 @@ var validMinimalUniversalConfig = `
       "CA": {
         "usages": [
           "cert sign",
-          "clr sign"
+          "crl sign"
         ],
         "expiry": "720h",
         "auth_key": "local-auth",
@@ -124,7 +122,7 @@ var validRemoteConfig = `
       "CA": {
         "usages": [
           "cert sign",
-          "clr sign"
+          "crl sign"
         ],
         "expiry": "720h",
         "auth_key": "ca-auth"
@@ -159,7 +157,7 @@ var validUniversalConfig = `
       "CA": {
         "usages": [
           "cert sign",
-          "clr sign"
+          "crl sign"
         ],
         "expiry": "720h",
         "auth_key": "local-auth",
@@ -212,7 +210,7 @@ var validNoAuthRemoteConfig = `
       "CA": {
         "usages": [
           "cert sign",
-          "clr sign"
+          "crl sign"
         ],
         "expiry": "720h"
       },
@@ -240,7 +238,7 @@ var validNoAuthUniversalConfig = `
       "CA": {
         "usages": [
           "cert sign",
-          "clr sign"
+          "crl sign"
         ],
         "expiry": "720h",
 		"remote": "localhost"
@@ -289,13 +287,37 @@ func TestNewSigner(t *testing.T) {
 	}
 }
 
+func checkInfo(t *testing.T, s signer.Signer, name string, profile *config.SigningProfile) {
+	req := info.Req{
+		Profile: name,
+	}
+	resp, err := s.Info(req)
+	if err != nil {
+		t.Fatal("remote info failed:", err)
+	}
+
+	if strings.Join(profile.Usage, ",") != strings.Join(resp.Usage, ",") {
+		t.Fatalf("Expected usage for profile %s to be %+v, got %+v", name, profile.Usage, resp.Usage)
+	}
+
+	caBytes, err := ioutil.ReadFile(testCaFile)
+	caBytes = bytes.TrimSpace(caBytes)
+	if err != nil {
+		t.Fatal("fail to read test CA cert:", err)
+	}
+
+	if bytes.Compare(caBytes, []byte(resp.Certificate)) != 0 {
+		t.Fatal("Get a different CA cert through info api.", len(resp.Certificate), len(caBytes))
+	}
+}
+
 func TestUniversalRemoteAndLocalInfo(t *testing.T) {
 	// set up remote server
-	remoteConfig := newConfig(t, []byte(validMinimalRemoteConfig))
+	remoteConfig := testsuite.NewConfig(t, []byte(validMinimalRemoteConfig))
 	remoteServer := newTestInfoServer(t, newTestUniversalSigner(t, remoteConfig.Signing))
 	defer closeTestServer(t, remoteServer)
 
-	universalConfig := newConfig(t, []byte(validMinimalUniversalConfig))
+	universalConfig := testsuite.NewConfig(t, []byte(validMinimalUniversalConfig))
 	// override with test server address, ignore url prefix "http://"
 	for name, profile := range universalConfig.Signing.Profiles {
 		if profile.RemoteServer != "" {
@@ -305,37 +327,20 @@ func TestUniversalRemoteAndLocalInfo(t *testing.T) {
 	s := newTestUniversalSigner(t, universalConfig.Signing)
 
 	for name, profile := range universalConfig.Signing.Profiles {
-		req := info.Req{
-			Profile: name,
-		}
-		resp, err := s.Info(req)
-		if err != nil {
-			t.Fatal("remote info failed:", err)
-		}
-
-		if strings.Join(profile.Usage, ",") != strings.Join(resp.Usage, ",") {
-			t.Fatalf("Expected usage for profile %s to be %+v, got %+v", name, profile.Usage, resp.Usage)
-		}
-
-		caBytes, err := ioutil.ReadFile(testCaFile)
-		caBytes = bytes.TrimSpace(caBytes)
-		if err != nil {
-			t.Fatal("fail to read test CA cert:", err)
-		}
-
-		if bytes.Compare(caBytes, []byte(resp.Certificate)) != 0 {
-			t.Fatal("Get a different CA cert through info api.", len(resp.Certificate), len(caBytes))
-		}
+		checkInfo(t, s, name, profile)
 	}
+
+	// add check for default profile
+	checkInfo(t, s, "", universalConfig.Signing.Default)
 }
 
 func TestUniversalMultipleRemoteAndLocalInfo(t *testing.T) {
 	// set up remote server
-	remoteConfig := newConfig(t, []byte(validRemoteConfig))
+	remoteConfig := testsuite.NewConfig(t, []byte(validRemoteConfig))
 	remoteServer := newTestInfoServer(t, newTestUniversalSigner(t, remoteConfig.Signing))
 	defer closeTestServer(t, remoteServer)
 
-	universalConfig := newConfig(t, []byte(validUniversalConfig))
+	universalConfig := testsuite.NewConfig(t, []byte(validUniversalConfig))
 	// override with test server address, ignore url prefix "http://"
 	for name, profile := range universalConfig.Signing.Profiles {
 		if profile.RemoteServer != "" {
@@ -345,84 +350,20 @@ func TestUniversalMultipleRemoteAndLocalInfo(t *testing.T) {
 	s := newTestUniversalSigner(t, universalConfig.Signing)
 
 	for name, profile := range universalConfig.Signing.Profiles {
-		req := info.Req{
-			Profile: name,
-		}
-		resp, err := s.Info(req)
-		if err != nil {
-			t.Fatal("remote info failed:", err)
-		}
-
-		if strings.Join(profile.Usage, ",") != strings.Join(resp.Usage, ",") {
-			t.Fatalf("Expected usage for profile %s to be %+v, got %+v", name, profile.Usage, resp.Usage)
-		}
-
-		caBytes, err := ioutil.ReadFile(testCaFile)
-		caBytes = bytes.TrimSpace(caBytes)
-		if err != nil {
-			t.Fatal("fail to read test CA cert:", err)
-		}
-
-		if bytes.Compare(caBytes, []byte(resp.Certificate)) != 0 {
-			t.Fatal("Get a different CA cert through info api.", len(resp.Certificate), len(caBytes))
-		}
+		checkInfo(t, s, name, profile)
 	}
-}
 
-type csrTest struct {
-	file    string
-	keyAlgo string
-	keyLen  int
-	// Error checking function
-	errorCallback func(*testing.T, error)
-}
-
-var csrTests = []csrTest{
-	{
-		file:          "../local/testdata/rsa2048.csr",
-		keyAlgo:       "rsa",
-		keyLen:        2048,
-		errorCallback: nil,
-	},
-	{
-		file:          "../local/testdata/rsa3072.csr",
-		keyAlgo:       "rsa",
-		keyLen:        3072,
-		errorCallback: nil,
-	},
-	{
-		file:          "../local/testdata/rsa4096.csr",
-		keyAlgo:       "rsa",
-		keyLen:        4096,
-		errorCallback: nil,
-	},
-	{
-		file:          "../local/testdata/ecdsa256.csr",
-		keyAlgo:       "ecdsa",
-		keyLen:        256,
-		errorCallback: nil,
-	},
-	{
-		file:          "../local/testdata/ecdsa384.csr",
-		keyAlgo:       "ecdsa",
-		keyLen:        384,
-		errorCallback: nil,
-	},
-	{
-		file:          "../local/testdata/ecdsa521.csr",
-		keyAlgo:       "ecdsa",
-		keyLen:        521,
-		errorCallback: nil,
-	},
+	// add check for default profile
+	checkInfo(t, s, "", universalConfig.Signing.Default)
 }
 
 func TestUniversalRemoteAndLocalSign(t *testing.T) {
 	// set up remote server
-	remoteConfig := newConfig(t, []byte(validNoAuthRemoteConfig))
+	remoteConfig := testsuite.NewConfig(t, []byte(validNoAuthRemoteConfig))
 	remoteServer := newTestSignServer(t, newTestUniversalSigner(t, remoteConfig.Signing))
 	defer closeTestServer(t, remoteServer)
 
-	universalConfig := newConfig(t, []byte(validNoAuthUniversalConfig))
+	universalConfig := testsuite.NewConfig(t, []byte(validNoAuthUniversalConfig))
 	// override with test server address, ignore url prefix "http://"
 	for name, profile := range universalConfig.Signing.Profiles {
 		if profile.RemoteServer != "" {
@@ -431,10 +372,10 @@ func TestUniversalRemoteAndLocalSign(t *testing.T) {
 	}
 	s := newTestUniversalSigner(t, universalConfig.Signing)
 
-	for name, profile := range universalConfig.Signing.Profiles {
+	checkSign := func(name string, profile *config.SigningProfile) {
 		hosts := []string{"cloudflare.com"}
-		for _, test := range csrTests {
-			csr, err := ioutil.ReadFile(test.file)
+		for _, test := range testsuite.CSRTests {
+			csr, err := ioutil.ReadFile(test.File)
 			if err != nil {
 				t.Fatalf("CSR loading error (%s): %v", name, err)
 			}
@@ -446,11 +387,11 @@ func TestUniversalRemoteAndLocalSign(t *testing.T) {
 				Serial:  testSerial,
 				Profile: name,
 			})
-			if test.errorCallback != nil {
-				test.errorCallback(t, err)
+			if test.ErrorCallback != nil {
+				test.ErrorCallback(t, err)
 			} else {
 				if err != nil {
-					t.Fatalf("Expected no error. Got %s. Param %s %d", err.Error(), test.keyAlgo, test.keyLen)
+					t.Fatalf("Expected no error. Got %s. Param %s %d", err.Error(), test.KeyAlgo, test.KeyLen)
 				}
 				cert, err := helpers.ParseCertificatePEM(certBytes)
 				if err != nil {
@@ -463,17 +404,13 @@ func TestUniversalRemoteAndLocalSign(t *testing.T) {
 			}
 		}
 	}
-}
 
-func newConfig(t *testing.T, configBytes []byte) *config.Config {
-	conf, err := config.LoadConfig([]byte(configBytes))
-	if err != nil {
-		t.Fatal("config loading error:", err)
+	for name, profile := range universalConfig.Signing.Profiles {
+		checkSign(name, profile)
 	}
-	if !conf.Valid() {
-		t.Fatal("config is not valid")
-	}
-	return conf
+
+	// add check for default profile
+	checkSign("", universalConfig.Signing.Default)
 }
 
 func newTestUniversalSigner(t *testing.T, policy *config.Signing) signer.Signer {
@@ -496,7 +433,7 @@ func newTestUniversalSigner(t *testing.T, policy *config.Signing) signer.Signer 
 }
 
 func newTestSignHandler(t *testing.T, s signer.Signer) (h http.Handler) {
-	h, err := NewSignHandlerFromSigner(s)
+	h, err := apisign.NewHandlerFromSigner(s)
 	if err != nil {
 		t.Fatal(err)
 	}
@@ -532,100 +469,4 @@ func newTestInfoServer(t *testing.T, s signer.Signer) *httptest.Server {
 func closeTestServer(t *testing.T, ts *httptest.Server) {
 	t.Log("Finalizing test server.")
 	ts.Close()
-}
-
-// NewSignHandlerFromSigner generates a new SignHandler directly from
-// an existing signer.
-func NewSignHandlerFromSigner(s signer.Signer) (h http.Handler, err error) {
-	policy := s.Policy()
-	if policy == nil {
-		err = errors.New(errors.PolicyError, errors.InvalidPolicy)
-		return
-	}
-
-	// Sign will only respond for profiles that have no auth provider.
-	// So if all of the profiles require authentication, we return an error.
-	haveUnauth := (policy.Default.Provider == nil)
-	for _, profile := range policy.Profiles {
-		if !haveUnauth {
-			break
-		}
-		haveUnauth = (profile.Provider == nil)
-	}
-
-	if !haveUnauth {
-		err = errors.New(errors.PolicyError, errors.InvalidPolicy)
-		return
-	}
-
-	return &api.HTTPHandler{
-		Handler: &SignHandler{
-			signer: s,
-		},
-		Methods: []string{"POST"},
-	}, nil
-}
-
-// A SignHandler accepts requests with a hostname and certficate
-// parameter (which should be PEM-encoded) and returns a new signed
-// certificate. It includes upstream servers indexed by their
-// profile name.
-type SignHandler struct {
-	signer signer.Signer
-}
-
-// Handle responds to requests for the CA to sign the certificate request
-// present in the "certificate_request" parameter for the host named
-// in the "hostname" parameter. The certificate should be PEM-encoded. If
-// provided, subject information from the "subject" parameter will be used
-// in place of the subject information from the CSR.
-func (h *SignHandler) Handle(w http.ResponseWriter, r *http.Request) error {
-	log.Info("signature request received")
-
-	body, err := ioutil.ReadAll(r.Body)
-	if err != nil {
-		return err
-	}
-	r.Body.Close()
-
-	var req signer.SignRequest
-	err = json.Unmarshal(body, &req)
-	if err != nil {
-		return err
-	}
-
-	if len(req.Hosts) == 0 {
-		return errors.NewBadRequestString("missing paratmeter 'hosts'")
-	}
-
-	if req.Request == "" {
-		return errors.NewBadRequestString("missing parameter 'certificate_request'")
-	}
-
-	var cert []byte
-	var profile *config.SigningProfile
-
-	policy := h.signer.Policy()
-	if policy != nil && policy.Profiles != nil && req.Profile != "" {
-		profile = policy.Profiles[req.Profile]
-	}
-
-	if profile == nil && policy != nil {
-		profile = policy.Default
-	}
-
-	if profile.Provider != nil {
-		log.Error("profile requires authentication")
-		return errors.NewBadRequestString("authentication required")
-	}
-
-	cert, err = h.signer.Sign(req)
-	if err != nil {
-		log.Warningf("failed to sign request: %v", err)
-		return err
-	}
-
-	result := map[string]string{"certificate": string(cert)}
-	log.Info("wrote response")
-	return api.SendResponse(w, result)
 }
